+++ scrapegraphai/nodes/generate_answer_node.py
@@ -4,16 +4,18 @@ GenerateAnswerNode Module
 
 # Imports from standard library
 from typing import List, Optional
-from tqdm import tqdm
 
 # Imports from Langchain
 from langchain.prompts import PromptTemplate
-from langchain_core.output_parsers import JsonOutputParser
+from langchain_core.output_parsers import JsonOutputParser, PydanticOutputParser
 from langchain_core.runnables import RunnableParallel
+from tqdm import tqdm
 
+from ..utils.logging import get_logger
+from ..models import Ollama
 # Imports from the library
 from .base_node import BaseNode
-from ..helpers import template_chunks, template_no_chunks, template_merge, template_chunks_with_schema, template_no_chunks_with_schema
+from ..helpers import template_chunks, template_no_chunks, template_merge
 
 
 class GenerateAnswerNode(BaseNode):
@@ -34,14 +36,23 @@ class GenerateAnswerNode(BaseNode):
         node_name (str): The unique identifier name for the node, defaulting to "GenerateAnswer".
     """
 
-    def __init__(self, input: str, output: List[str], node_config: Optional[dict] = None,
-                 node_name: str = "GenerateAnswer"):
-
+    def __init__(
+        self,
+        input: str,
+        output: List[str],
+        node_config: Optional[dict] = None,
+        node_name: str = "GenerateAnswer",
+    ):
         super().__init__(node_name, "node", input, output, 2, node_config)
-
+      
         self.llm_model = node_config["llm_model"]
-        self.verbose = False if node_config is None else node_config.get(
-            "verbose", False)
+
+        if isinstance(node_config["llm_model"], Ollama):
+            self.llm_model.format="json"
+
+        self.verbose = (
+            True if node_config is None else node_config.get("verbose", False)
+        )
 
     def execute(self, state: dict) -> dict:
         """
@@ -60,8 +71,8 @@ class GenerateAnswerNode(BaseNode):
                       that the necessary information for generating an answer is missing.
         """
 
-        if self.verbose:
-            print(f"--- Executing {self.node_name} Node ---")
+        self.logger.info(f"--- Executing {self.node_name} Node ---")
+
         # Interpret input keys based on the provided input expression
         input_keys = self.get_input_keys(state)
         # Fetching data from the state based on the input keys
@@ -69,42 +80,32 @@ class GenerateAnswerNode(BaseNode):
         user_prompt = input_data[0]
         doc = input_data[1]
 
-        output_parser = JsonOutputParser()
+        # Initialize the output parser
+        if self.node_config.get("schema",None) is not None:
+            output_parser = PydanticOutputParser(pydantic_object=self.node_config.get("schema", None))
+        else:
+            output_parser = JsonOutputParser()
+
         format_instructions = output_parser.get_format_instructions()
 
         chains_dict = {}
 
         # Use tqdm to add progress bar
         for i, chunk in enumerate(tqdm(doc, desc="Processing chunks", disable=not self.verbose)):
-            if self.node_config["schema"] is None and len(doc) == 1:
+            if len(doc) == 1:
                 prompt = PromptTemplate(
                     template=template_no_chunks,
                     input_variables=["question"],
                     partial_variables={"context": chunk.page_content,
                                        "format_instructions": format_instructions})
-            elif self.node_config["schema"] is not None and len(doc) == 1:
-                 prompt = PromptTemplate(
-                    template=template_no_chunks_with_schema,
-                    input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                       "format_instructions": format_instructions,
-                                       "schema": self.node_config["schema"]
-                                       })
-            elif self.node_config["schema"] is None and len(doc) > 1:
+                
+            else:
                 prompt = PromptTemplate(
                     template=template_chunks,
                     input_variables=["question"],
                     partial_variables={"context": chunk.page_content,
                                         "chunk_id": i + 1,
                                         "format_instructions": format_instructions})
-            elif self.node_config["schema"] is not None and len(doc) > 1:
-                prompt = PromptTemplate(
-                    template=template_chunks_with_schema,
-                    input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                        "chunk_id": i + 1,
-                                        "format_instructions": format_instructions,
-                                        "schema": self.node_config["schema"]})
 
             # Dynamically name the chains based on their index
             chain_name = f"chunk{i+1}"
@@ -122,12 +123,14 @@ class GenerateAnswerNode(BaseNode):
                 partial_variables={"format_instructions": format_instructions},
             )
             merge_chain = merge_prompt | self.llm_model | output_parser
-            answer = merge_chain.invoke(
-                {"context": answer, "question": user_prompt})
+            answer = merge_chain.invoke({"context": answer, "question": user_prompt})
         else:
             # Chain
             single_chain = list(chains_dict.values())[0]
             answer = single_chain.invoke({"question": user_prompt})
+
+        if type(answer) == PydanticOutputParser:
+            answer = answer.model_dump()
 
         # Update the state with the generated answer
         state.update({self.output[0]: answer})