+++ scrapegraphai/nodes/generate_answer_omni_node.py
@@ -4,13 +4,13 @@ GenerateAnswerNode Module
 
 # Imports from standard library
 from typing import List, Optional
-from tqdm import tqdm
 
 # Imports from Langchain
 from langchain.prompts import PromptTemplate
-from langchain_core.output_parsers import JsonOutputParser
+from langchain_core.output_parsers import JsonOutputParser, PydanticOutputParser
 from langchain_core.runnables import RunnableParallel
-
+from tqdm import tqdm
+from ..models import Ollama
 # Imports from the library
 from .base_node import BaseNode
 from ..helpers.generate_answer_node_omni_prompts import template_no_chunk_omni, template_chunks_omni, template_merge_omni
@@ -34,13 +34,22 @@ class GenerateAnswerOmniNode(BaseNode):
         node_name (str): The unique identifier name for the node, defaulting to "GenerateAnswer".
     """
 
-    def __init__(self, input: str, output: List[str], node_config: Optional[dict] = None,
-                 node_name: str = "GenerateAnswerOmni"):
+    def __init__(
+        self,
+        input: str,
+        output: List[str],
+        node_config: Optional[dict] = None,
+        node_name: str = "GenerateAnswerOmni",
+    ):
         super().__init__(node_name, "node", input, output, 3, node_config)
 
         self.llm_model = node_config["llm_model"]
-        self.verbose = False if node_config is None else node_config.get(
-            "verbose", False)
+        if isinstance(node_config["llm_model"], Ollama):
+            self.llm_model.format="json"
+            
+        self.verbose = (
+            False if node_config is None else node_config.get("verbose", False)
+        )
 
     def execute(self, state: dict) -> dict:
         """
@@ -59,8 +68,7 @@ class GenerateAnswerOmniNode(BaseNode):
                       that the necessary information for generating an answer is missing.
         """
 
-        if self.verbose:
-            print(f"--- Executing {self.node_name} Node ---")
+        self.logger.info(f"--- Executing {self.node_name} Node ---")
 
         # Interpret input keys based on the provided input expression
         input_keys = self.get_input_keys(state)
@@ -72,29 +80,40 @@ class GenerateAnswerOmniNode(BaseNode):
         doc = input_data[1]
         imag_desc = input_data[2]
 
-        output_parser = JsonOutputParser()
+        # Initialize the output parser
+        if self.node_config.get("schema", None) is not None:
+            output_parser = PydanticOutputParser(pydantic_object=self.node_config.get("schema", None))
+        else:
+            output_parser = JsonOutputParser()
+
         format_instructions = output_parser.get_format_instructions()
 
 
         chains_dict = {}
 
         # Use tqdm to add progress bar
-        for i, chunk in enumerate(tqdm(doc, desc="Processing chunks", disable=not self.verbose)):
+        for i, chunk in enumerate(
+            tqdm(doc, desc="Processing chunks", disable=not self.verbose)
+        ):
             if len(doc) == 1:
                 prompt = PromptTemplate(
                     template=template_no_chunk_omni,
                     input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                       "format_instructions": format_instructions,
-                                        "img_desc": imag_desc},
+                    partial_variables={
+                        "context": chunk.page_content,
+                        "format_instructions": format_instructions,
+                        "img_desc": imag_desc,
+                    },
                 )
             else:
                 prompt = PromptTemplate(
                     template=template_chunks_omni,
                     input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                       "chunk_id": i + 1,
-                                       "format_instructions": format_instructions},
+                    partial_variables={
+                        "context": chunk.page_content,
+                        "chunk_id": i + 1,
+                        "format_instructions": format_instructions,
+                    },
                 )
 
             # Dynamically name the chains based on their index
@@ -116,12 +135,14 @@ class GenerateAnswerOmniNode(BaseNode):
                 },
             )
             merge_chain = merge_prompt | self.llm_model | output_parser
-            answer = merge_chain.invoke(
-                {"context": answer, "question": user_prompt})
+            answer = merge_chain.invoke({"context": answer, "question": user_prompt})
         else:
             # Chain
             single_chain = list(chains_dict.values())[0]
             answer = single_chain.invoke({"question": user_prompt})
+
+        if type(answer) == PydanticOutputParser:
+            answer = answer.model_dump()
 
         # Update the state with the generated answer
         state.update({self.output[0]: answer})