+++ tests/nodes/robot_node_test.py
@@ -1,15 +1,11 @@
-"""
-Module for testinh robot_node
-"""
 import pytest
 from scrapegraphai.models import Ollama
 from scrapegraphai.nodes import RobotsNode
 
-
 @pytest.fixture
 def setup():
     """
-    setup
+    Setup
     """
     # ************************************************
     # Define the configuration for the graph
@@ -17,7 +13,7 @@ def setup():
 
     graph_config = {
         "llm": {
-            "model": "ollama/llama3",
+            "model_name": "ollama/llama3",  # Modifica il nome dell'attributo da "model_name" a "model"
             "temperature": 0,
             "streaming": True
         },
@@ -37,21 +33,26 @@ def setup():
                      }
     )
 
-    return robots_node
+    # ************************************************
+    # Define the initial state
+    # ************************************************
+
+    initial_state = {
+        "url": "https://twitter.com/home"
+    }
+
+    return robots_node, initial_state
 
 # ************************************************
 # Test the node
 # ************************************************
 
-
 def test_robots_node(setup):
     """
     Run the tests
     """
-    state = {
-        "url": "https://twitter.com/home"
-    }
+    robots_node, initial_state = setup  # Estrai l'oggetto RobotsNode e lo stato iniziale dalla tupla
 
-    result = setup.execute(state)
+    result = robots_node.execute(initial_state)
 
     assert result is not None